Refactor ExtFile to brut.j.dir
 - move to using Paths when possible
 - adapt calls to FileDirectory to leverage Paths
