+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/AndResGuardTest.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib;
 
-import brut.androlib.res.util.ExtFile;
+import brut.directory.ExtFile;
 import brut.common.BrutException;
 import brut.util.OS;
 import java.io.File;
