+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/DebugTagRetainedTest.java
@@ -15,7 +15,7 @@
  */
 package brut.androlib;
 
-import brut.androlib.res.util.ExtFile;
+import brut.directory.ExtFile;
 import brut.common.BrutException;
 import brut.util.OS;
 import org.junit.AfterClass;
