+++ brut.j.dir/src/main/java/brut/directory/ExtFile.java
@@ -14,12 +14,8 @@
  *  limitations under the License.
  */
 
-package brut.androlib.res.util;
+package brut.directory;
 
-import brut.directory.Directory;
-import brut.directory.DirectoryException;
-import brut.directory.FileDirectory;
-import brut.directory.ZipRODirectory;
 import java.io.File;
 import java.net.URI;
